+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -7,7 +7,6 @@ import io.sentry.hints.ApplyScopeData;
 import io.sentry.hints.Backfillable;
 import io.sentry.hints.Cached;
 import io.sentry.hints.DiskFlushNotification;
-import io.sentry.hints.EventDropReason;
 import io.sentry.hints.TransactionEnd;
 import io.sentry.logger.ILoggerBatchProcessor;
 import io.sentry.logger.NoOpLoggerBatchProcessor;
@@ -138,7 +137,6 @@ public final class SentryClient implements ISentryClient {
         options
             .getClientReportRecorder()
             .recordLostEvent(DiscardReason.EVENT_PROCESSOR, DataCategory.Error);
-        HintUtils.setEventDropReason(hint, EventDropReason.IGNORED);
         return SentryId.EMPTY_ID;
       }
 
@@ -152,7 +150,6 @@ public final class SentryClient implements ISentryClient {
         options
             .getClientReportRecorder()
             .recordLostEvent(DiscardReason.EVENT_PROCESSOR, DataCategory.Error);
-        HintUtils.setEventDropReason(hint, EventDropReason.IGNORED);
         return SentryId.EMPTY_ID;
       }
     }
@@ -179,7 +176,6 @@ public final class SentryClient implements ISentryClient {
         options
             .getClientReportRecorder()
             .recordLostEvent(DiscardReason.BEFORE_SEND, DataCategory.Error);
-        HintUtils.setEventDropReason(hint, EventDropReason.BEFORE_SEND);
       }
     }
 
@@ -212,7 +208,6 @@ public final class SentryClient implements ISentryClient {
         options
             .getClientReportRecorder()
             .recordLostEvent(DiscardReason.SAMPLE_RATE, DataCategory.Error);
-        HintUtils.setEventDropReason(hint, EventDropReason.SAMPLE_RATE);
         // setting event as null to not be sent as its been discarded by sample rate
         event = null;
       }
@@ -292,6 +287,7 @@ public final class SentryClient implements ISentryClient {
       options.getLogger().log(SentryLevel.WARNING, e, "Capturing event %s failed.", sentryId);
 
       // if there was an error capturing the event, we return an emptyId
+      HintUtils.setCaptureFailed(hint);
       sentryId = SentryId.EMPTY_ID;
     }